Break core code into modules
