+++ .github/workflows/rust.yml
@@ -16,3 +16,9 @@ jobs:
       run: cargo build --verbose
     - name: Run tests
       run: cargo test --verbose
+    - name: Run tests with zeroize
+      run: cargo test --verbose --features zeroize
+    - name: Run tests with use-unsafe
+      run: cargo test --verbose --features use-unsafe
+    - name: Run tests with all features
+      run: cargo test --verbose --all-features
