+++ src/fixeduint.rs
@@ -12,26 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use num_traits::ops::overflowing::{OverflowingAdd, OverflowingMul, OverflowingSub};
+use num_traits::ops::overflowing::{OverflowingAdd, OverflowingSub};
 use num_traits::{Bounded, One, PrimInt, ToPrimitive, Zero};
 
-use num_integer;
-
-#[cfg(feature = "use-unsafe")]
-use core::borrow::{Borrow, BorrowMut};
 use core::convert::TryFrom;
 use core::fmt::Write;
-#[cfg(feature = "use-unsafe")]
-use core::hash::Hash;
 
 use crate::machineword::MachineWord;
-use crate::patch_num_traits::{OverflowingShl, OverflowingShr};
-
-#[cfg(feature = "zeroize")]
-use zeroize::DefaultIsZeroes;
 
 #[allow(unused_imports)]
 use num_traits::{FromPrimitive, Num};
+
+mod add_sub_impl;
+mod bit_ops_impl;
+mod mul_div_impl;
+mod num_integer_impl;
+mod num_traits_casts;
+mod num_traits_identity;
+mod prim_int_impl;
+mod string_conversion;
+mod to_from_bytes;
 
 /// Fixed-size unsigned integer, represented by array of N words of builtin unsigned type T
 #[derive(Debug, Clone, Copy, core::cmp::PartialEq, core::cmp::Eq)]
@@ -528,631 +528,6 @@ impl<T: MachineWord, const N: usize> core::convert::From<u64> for FixedUInt<T, N
 
 // #endregion core::convert::From<primitive>
 
-// #region Add/Subtract
-
-impl<T: MachineWord, const N: usize> num_traits::ops::overflowing::OverflowingAdd
-    for FixedUInt<T, N>
-{
-    fn overflowing_add(&self, other: &Self) -> (Self, bool) {
-        let mut ret = *self;
-        let overflow = Self::add_impl(&mut ret, other);
-        (ret, overflow)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Add for FixedUInt<T, N> {
-    type Output = Self;
-    fn add(self, other: Self) -> Self {
-        let (res, overflow) = self.overflowing_add(&other);
-        if overflow {
-            maybe_panic(PanicReason::Add);
-        }
-        res
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Add<&'_ Self> for FixedUInt<T, N> {
-    type Output = Self;
-    fn add(self, other: &Self) -> Self {
-        let (res, overflow) = self.overflowing_add(other);
-        if overflow {
-            maybe_panic(PanicReason::Add);
-        }
-        res
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::WrappingAdd for FixedUInt<T, N> {
-    fn wrapping_add(&self, other: &Self) -> Self {
-        self.overflowing_add(other).0
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::CheckedAdd for FixedUInt<T, N> {
-    fn checked_add(&self, other: &Self) -> Option<Self> {
-        let res = self.overflowing_add(other);
-        if res.1 {
-            None
-        } else {
-            Some(res.0)
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::ops::saturating::SaturatingAdd
-    for FixedUInt<T, N>
-{
-    /// Saturating addition operator. Returns a+b, saturating at the numeric bounds instead of overflowing.
-    fn saturating_add(&self, other: &Self) -> Self {
-        self.saturating_add_impl(other)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::AddAssign<Self> for FixedUInt<T, N> {
-    fn add_assign(&mut self, other: Self) {
-        if Self::add_impl(self, &other) {
-            maybe_panic(PanicReason::Add);
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::AddAssign<&'_ Self> for FixedUInt<T, N> {
-    fn add_assign(&mut self, other: &Self) {
-        if Self::add_impl(self, other) {
-            maybe_panic(PanicReason::Add);
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::ops::overflowing::OverflowingSub
-    for FixedUInt<T, N>
-{
-    fn overflowing_sub(&self, other: &Self) -> (Self, bool) {
-        let mut ret = *self;
-        let overflow = Self::sub_impl(&mut ret, other);
-        (ret, overflow)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Sub for FixedUInt<T, N> {
-    type Output = Self;
-    fn sub(self, other: Self) -> <Self as core::ops::Sub<Self>>::Output {
-        let (res, overflow) = self.overflowing_sub(&other);
-        if overflow {
-            maybe_panic(PanicReason::Sub);
-        }
-        res
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Sub<&'_ Self> for FixedUInt<T, N> {
-    type Output = Self;
-    fn sub(self, other: &Self) -> Self {
-        let (res, overflow) = self.overflowing_sub(other);
-        if overflow {
-            maybe_panic(PanicReason::Sub);
-        }
-        res
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::WrappingSub for FixedUInt<T, N> {
-    fn wrapping_sub(&self, other: &Self) -> Self {
-        self.overflowing_sub(other).0
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::CheckedSub for FixedUInt<T, N> {
-    fn checked_sub(&self, other: &Self) -> Option<Self> {
-        let res = self.overflowing_sub(other);
-        if res.1 {
-            None
-        } else {
-            Some(res.0)
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::ops::saturating::SaturatingSub
-    for FixedUInt<T, N>
-{
-    /// Saturating subtraction operator. Returns a-b, saturating at the numeric bounds instead of overflowing.
-    fn saturating_sub(&self, other: &Self) -> Self {
-        self.saturating_sub_impl(other)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::SubAssign<Self> for FixedUInt<T, N> {
-    fn sub_assign(&mut self, other: Self) {
-        if Self::sub_impl(self, &other) {
-            maybe_panic(PanicReason::Sub);
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::SubAssign<&'_ Self> for FixedUInt<T, N> {
-    fn sub_assign(&mut self, other: &Self) {
-        if Self::sub_impl(self, other) {
-            maybe_panic(PanicReason::Sub);
-        }
-    }
-}
-
-/// Note: This is marked deprecated, but still used by PrimInt
-impl<T: MachineWord, const N: usize> num_traits::Saturating for FixedUInt<T, N> {
-    /// Saturating addition operator. Returns a+b, saturating at the numeric bounds instead of overflowing.
-    fn saturating_add(self, other: Self) -> Self {
-        self.saturating_add_impl(&other)
-    }
-
-    /// Saturating subtraction operator. Returns a-b, saturating at the numeric bounds instead of overflowing.
-    fn saturating_sub(self, other: Self) -> Self {
-        self.saturating_sub_impl(&other)
-    }
-}
-
-// #endregion Add/Subtract
-
-// #region Multiply/Divide
-
-impl<T: MachineWord, const N: usize> num_traits::ops::overflowing::OverflowingMul
-    for FixedUInt<T, N>
-{
-    fn overflowing_mul(&self, other: &Self) -> (Self, bool) {
-        Self::mul_impl::<true>(self, other)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Mul for FixedUInt<T, N> {
-    type Output = Self;
-    fn mul(self, other: Self) -> <Self as core::ops::Mul<Self>>::Output {
-        let res = self.overflowing_mul(&other);
-        if res.1 {
-            maybe_panic(PanicReason::Mul);
-        }
-        res.0
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Mul<&'_ Self> for FixedUInt<T, N> {
-    type Output = Self;
-    fn mul(self, other: &Self) -> <Self as core::ops::Mul<Self>>::Output {
-        let res = self.overflowing_mul(other);
-        if res.1 {
-            maybe_panic(PanicReason::Mul);
-        }
-        res.0
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::WrappingMul for FixedUInt<T, N> {
-    fn wrapping_mul(&self, other: &Self) -> Self {
-        Self::mul_impl::<false>(self, other).0
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::CheckedMul for FixedUInt<T, N> {
-    fn checked_mul(&self, other: &Self) -> Option<Self> {
-        let res = self.overflowing_mul(other);
-        if res.1 {
-            None
-        } else {
-            Some(res.0)
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::ops::saturating::SaturatingMul
-    for FixedUInt<T, N>
-{
-    fn saturating_mul(&self, other: &Self) -> Self {
-        let res = self.overflowing_mul(other);
-        if res.1 {
-            Self::max_value()
-        } else {
-            res.0
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::MulAssign<Self> for FixedUInt<T, N> {
-    fn mul_assign(&mut self, other: Self) {
-        let res = self.overflowing_mul(&other);
-        *self = res.0;
-        if res.1 {
-            maybe_panic(PanicReason::Mul);
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::MulAssign<&'_ Self> for FixedUInt<T, N> {
-    fn mul_assign(&mut self, other: &Self) {
-        let res = self.overflowing_mul(other);
-        *self = res.0;
-        if res.1 {
-            maybe_panic(PanicReason::Mul);
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Div for FixedUInt<T, N> {
-    type Output = Self;
-    fn div(self, other: Self) -> <Self as core::ops::Div<Self>>::Output {
-        if other.is_zero() {
-            maybe_panic(PanicReason::DivByZero)
-        }
-        Self::div_impl(&self, &other)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Div<&'_ Self> for FixedUInt<T, N> {
-    type Output = Self;
-    fn div(self, other: &Self) -> <Self as core::ops::Div<Self>>::Output {
-        if other.is_zero() {
-            maybe_panic(PanicReason::DivByZero)
-        }
-        Self::div_impl(&self, other)
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::CheckedDiv for FixedUInt<T, N> {
-    fn checked_div(&self, other: &Self) -> Option<Self> {
-        if other.is_zero() {
-            None
-        } else {
-            Some(core::ops::Div::<Self>::div(*self, *other))
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::DivAssign<Self> for FixedUInt<T, N> {
-    fn div_assign(&mut self, other: Self) {
-        if other.is_zero() {
-            maybe_panic(PanicReason::DivByZero)
-        }
-        *self = Self::div_impl(self, &other);
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::DivAssign<&'_ Self> for FixedUInt<T, N> {
-    fn div_assign(&mut self, other: &Self) {
-        if other.is_zero() {
-            maybe_panic(PanicReason::DivByZero)
-        }
-        *self = Self::div_impl(self, other);
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Rem for FixedUInt<T, N> {
-    type Output = Self;
-    fn rem(self, other: Self) -> Self {
-        if other.is_zero() {
-            maybe_panic(PanicReason::RemByZero)
-        }
-        self.div_rem(&other).1
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Rem<&'_ Self> for FixedUInt<T, N> {
-    type Output = Self;
-    fn rem(self, other: &Self) -> Self {
-        if other.is_zero() {
-            maybe_panic(PanicReason::RemByZero)
-        }
-        self.div_rem(other).1
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::CheckedRem for FixedUInt<T, N> {
-    fn checked_rem(&self, other: &Self) -> Option<Self> {
-        if other.is_zero() {
-            None
-        } else {
-            Some(core::ops::Rem::<Self>::rem(*self, *other))
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::RemAssign<Self> for FixedUInt<T, N> {
-    fn rem_assign(&mut self, other: Self) {
-        if other.is_zero() {
-            maybe_panic(PanicReason::RemByZero)
-        }
-        *self = self.div_rem(&other).1;
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::RemAssign<&'_ Self> for FixedUInt<T, N> {
-    fn rem_assign(&mut self, other: &Self) {
-        if other.is_zero() {
-            maybe_panic(PanicReason::RemByZero)
-        }
-        *self = self.div_rem(other).1;
-    }
-}
-
-// #endregion Multiply/Divide
-
-// #region Bitops
-
-impl<T: MachineWord, const N: usize> core::ops::Not for FixedUInt<T, N> {
-    type Output = Self;
-    fn not(self) -> <Self as core::ops::Not>::Output {
-        let mut ret = Self::zero();
-        for i in 0..N {
-            ret.array[i] = !self.array[i]
-        }
-        ret
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::BitAnd for FixedUInt<T, N> {
-    type Output = Self;
-    fn bitand(self, other: Self) -> <Self as core::ops::BitAnd<Self>>::Output {
-        let mut ret = Self::zero();
-        for i in 0..N {
-            ret.array[i] = self.array[i] & other.array[i]
-        }
-        ret
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::BitAndAssign for FixedUInt<T, N> {
-    fn bitand_assign(&mut self, other: Self) {
-        for i in 0..N {
-            self.array[i] &= other.array[i]
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::BitOr for FixedUInt<T, N> {
-    type Output = Self;
-    fn bitor(self, other: Self) -> <Self as core::ops::BitOr<Self>>::Output {
-        let mut ret = Self::zero();
-        for i in 0..N {
-            ret.array[i] = self.array[i] | other.array[i]
-        }
-        ret
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::BitOrAssign for FixedUInt<T, N> {
-    fn bitor_assign(&mut self, other: Self) {
-        for i in 0..N {
-            self.array[i] |= other.array[i]
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::BitXor for FixedUInt<T, N> {
-    type Output = Self;
-    fn bitxor(self, other: Self) -> <Self as core::ops::BitXor<Self>>::Output {
-        let mut ret = Self::zero();
-        for i in 0..N {
-            ret.array[i] = self.array[i] ^ other.array[i]
-        }
-        ret
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::BitXorAssign for FixedUInt<T, N> {
-    fn bitxor_assign(&mut self, other: Self) {
-        for i in 0..N {
-            self.array[i] ^= other.array[i]
-        }
-    }
-}
-
-impl<T: MachineWord, const N: usize> OverflowingShl for FixedUInt<T, N> {
-    fn overflowing_shl(self, bits: u32) -> (Self, bool) {
-        let bitsu = bits as usize;
-        let (shift, overflow) = if bitsu >= Self::BIT_SIZE {
-            (bitsu & (Self::BIT_SIZE - 1), true)
-        } else {
-            (bitsu, false)
-        };
-        let res = core::ops::Shl::<usize>::shl(self, shift);
-        (res, overflow)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Shl<u32> for FixedUInt<T, N> {
-    type Output = Self;
-    fn shl(self, bits: u32) -> <Self as core::ops::Shl<u32>>::Output {
-        core::ops::Shl::<usize>::shl(self, bits as usize)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Shl<usize> for FixedUInt<T, N> {
-    type Output = Self;
-    fn shl(self, bits: usize) -> <Self as core::ops::Shl<usize>>::Output {
-        // This copy can be avoided
-        let mut result = self;
-        Self::shl_impl(&mut result, bits);
-        result
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Shl<&'_ usize> for FixedUInt<T, N> {
-    type Output = Self;
-    fn shl(self, bits: &usize) -> <Self as core::ops::Shl<usize>>::Output {
-        let mut result = self;
-        Self::shl_impl(&mut result, *bits);
-        result
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::WrappingShl for FixedUInt<T, N> {
-    fn wrapping_shl(&self, bits: u32) -> Self {
-        self.overflowing_shl(bits).0
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::CheckedShl for FixedUInt<T, N> {
-    fn checked_shl(&self, bits: u32) -> Option<Self> {
-        let res = self.overflowing_shl(bits);
-        if res.1 {
-            None
-        } else {
-            Some(res.0)
-        }
-    }
-}
-
-// SaturatingShl doesn't exist
-
-impl<T: MachineWord, const N: usize> core::ops::ShlAssign<usize> for FixedUInt<T, N> {
-    fn shl_assign(&mut self, bits: usize) {
-        Self::shl_impl(self, bits);
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::ShlAssign<&'_ usize> for FixedUInt<T, N> {
-    fn shl_assign(&mut self, bits: &usize) {
-        Self::shl_impl(self, *bits);
-    }
-}
-
-impl<T: MachineWord, const N: usize> OverflowingShr for FixedUInt<T, N> {
-    fn overflowing_shr(self, bits: u32) -> (Self, bool) {
-        let bitsu = bits as usize;
-        let (shift, overflow) = if bitsu >= Self::BIT_SIZE {
-            (bitsu & (Self::BIT_SIZE - 1), true)
-        } else {
-            (bitsu, false)
-        };
-        let res = core::ops::Shr::<usize>::shr(self, shift);
-        (res, overflow)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Shr<u32> for FixedUInt<T, N> {
-    type Output = Self;
-    fn shr(self, bits: u32) -> <Self as core::ops::Shr<u32>>::Output {
-        core::ops::Shr::<usize>::shr(self, bits as usize)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Shr<usize> for FixedUInt<T, N> {
-    type Output = Self;
-    fn shr(self, bits: usize) -> <Self as core::ops::Shr<usize>>::Output {
-        // Technically, this copy can be avoided
-        let mut result = self;
-        Self::shr_impl(&mut result, bits);
-        result
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::Shr<&'_ usize> for FixedUInt<T, N> {
-    type Output = Self;
-    fn shr(self, bits: &usize) -> <Self as core::ops::Shr<usize>>::Output {
-        let mut result = self;
-        Self::shr_impl(&mut result, *bits);
-        result
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::WrappingShr for FixedUInt<T, N> {
-    fn wrapping_shr(&self, bits: u32) -> Self {
-        self.overflowing_shr(bits).0
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::CheckedShr for FixedUInt<T, N> {
-    fn checked_shr(&self, bits: u32) -> Option<Self> {
-        let res = self.overflowing_shr(bits);
-        if res.1 {
-            None
-        } else {
-            Some(res.0)
-        }
-    }
-}
-
-// SaturatingShr doesn't exist
-
-impl<T: MachineWord, const N: usize> core::ops::ShrAssign<usize> for FixedUInt<T, N> {
-    fn shr_assign(&mut self, bits: usize) {
-        Self::shr_impl(self, bits);
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::ops::ShrAssign<&'_ usize> for FixedUInt<T, N> {
-    fn shr_assign(&mut self, bits: &usize) {
-        Self::shr_impl(self, *bits);
-    }
-}
-
-// #endregion Bitops
-
-// #region num_traits Identity
-
-impl<T: MachineWord, const N: usize> num_traits::Zero for FixedUInt<T, N> {
-    fn zero() -> Self {
-        Self::new()
-    }
-    fn is_zero(&self) -> bool {
-        !self.array.iter().any(|v| !v.is_zero())
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::One for FixedUInt<T, N> {
-    fn one() -> Self {
-        let mut ret = Self::zero();
-        ret.array[0] = T::one();
-        ret
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::Bounded for FixedUInt<T, N> {
-    fn min_value() -> Self {
-        Self::zero()
-    }
-    fn max_value() -> Self {
-        FixedUInt {
-            array: [T::max_value(); N],
-        }
-    }
-}
-
-// #endregion num_traits Identity
-
-// #region num_traits casting
-
-impl<T: MachineWord, const N: usize> num_traits::NumCast for FixedUInt<T, N> {
-    fn from<X>(_: X) -> Option<Self>
-    where
-        T: ToPrimitive,
-    {
-        todo!()
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::ToPrimitive for FixedUInt<T, N> {
-    fn to_i64(&self) -> Option<i64> {
-        None
-    }
-    fn to_u64(&self) -> Option<u64> {
-        let mut ret: u64 = 0;
-        let iter: usize = core::cmp::min(8 / Self::WORD_SIZE, N);
-        for (word, bits) in (0..iter).map(|x| (x, x as u64 * Self::WORD_SIZE as u64 * 8)) {
-            ret += T::to_u64(&self.array[word])? << bits;
-        }
-        Some(ret)
-    }
-}
-
-impl<T: MachineWord, const N: usize> num_traits::FromPrimitive for FixedUInt<T, N> {
-    fn from_i64(_: i64) -> Option<Self> {
-        None
-    }
-    fn from_u64(input: u64) -> Option<Self> {
-        Some(Self::from_le_bytes(&input.to_le_bytes()))
-    }
-}
-
-// #endregion num_traits casting
-
 // #region helpers
 
 // This is slightly less than ideal, but PIE isn't directly constructible
@@ -1208,343 +583,6 @@ fn maybe_panic(r: PanicReason) {
 }
 
 // #endregion helpers
-
-// #region strings
-
-impl<T: MachineWord, const N: usize> num_traits::Num for FixedUInt<T, N> {
-    type FromStrRadixErr = core::num::ParseIntError;
-    fn from_str_radix(
-        input: &str,
-        radix: u32,
-    ) -> Result<Self, <Self as num_traits::Num>::FromStrRadixErr> {
-        if input.is_empty() {
-            return Err(make_empty_error());
-        }
-
-        if !(2..=16).contains(&radix) {
-            return Err(make_overflow_err()); // Invalid radix
-        }
-
-        let mut ret = Self::zero();
-        let range = match input.find(|c: char| c != '0') {
-            Some(x) => &input[x..],
-            _ => input,
-        };
-
-        for c in range.chars() {
-            let digit = match c.to_digit(radix) {
-                Some(d) => d,
-                None => return Err(make_parse_int_err()), // Invalid character for the radix
-            };
-
-            ret = num_traits::CheckedMul::checked_mul(&ret, &Self::from(radix as u8))
-                .ok_or(make_overflow_err())?;
-            ret = num_traits::CheckedAdd::checked_add(&ret, &Self::from(digit as u8))
-                .ok_or(make_overflow_err())?;
-        }
-
-        Ok(ret)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::fmt::UpperHex for FixedUInt<T, N>
-where
-    u8: core::convert::TryFrom<T>,
-{
-    fn fmt(&self, formatter: &mut core::fmt::Formatter<'_>) -> Result<(), core::fmt::Error> {
-        self.hex_fmt(formatter, true)
-    }
-}
-
-impl<T: MachineWord, const N: usize> core::fmt::LowerHex for FixedUInt<T, N>
-where
-    u8: core::convert::TryFrom<T>,
-{
-    fn fmt(&self, formatter: &mut core::fmt::Formatter<'_>) -> Result<(), core::fmt::Error> {
-        self.hex_fmt(formatter, false)
-    }
-}
-
-// #endregion strings
-
-// #region unimplemented
-
-impl<T: MachineWord, const N: usize> num_traits::PrimInt for FixedUInt<T, N> {
-    fn count_ones(self) -> u32 {
-        self.array.iter().map(|&val| val.count_ones()).sum()
-    }
-    fn count_zeros(self) -> u32 {
-        self.array.iter().map(|&val| val.count_zeros()).sum()
-    }
-    fn leading_zeros(self) -> u32 {
-        let mut ret = 0u32;
-        for index in (0..N).rev() {
-            let v = self.array[index];
-            ret += v.leading_zeros();
-            if !v.is_zero() {
-                break;
-            }
-        }
-        ret
-    }
-    fn trailing_zeros(self) -> u32 {
-        let mut ret = 0u32;
-        for index in 0..N {
-            let v = self.array[index];
-            ret += v.trailing_zeros();
-            if !v.is_zero() {
-                break;
-            }
-        }
-        ret
-    }
-    fn rotate_left(self, bits: u32) -> Self {
-        let a = self << bits;
-        let b = self >> (Self::BIT_SIZE - bits as usize);
-        a | b
-    }
-    fn rotate_right(self, bits: u32) -> Self {
-        let a = self >> bits;
-        let b = self << (Self::BIT_SIZE - bits as usize);
-        a | b
-    }
-    fn signed_shl(self, _: u32) -> Self {
-        todo!()
-    }
-    fn signed_shr(self, _: u32) -> Self {
-        todo!()
-    }
-    fn unsigned_shl(self, bits: u32) -> Self {
-        core::ops::Shl::<u32>::shl(self, bits)
-    }
-    fn unsigned_shr(self, bits: u32) -> Self {
-        core::ops::Shr::<u32>::shr(self, bits)
-    }
-    fn swap_bytes(self) -> Self {
-        let mut ret = Self::new();
-        for index in 0..N {
-            ret.array[index] = self.array[N - 1 - index].swap_bytes();
-        }
-
-        ret
-    }
-    fn from_be(source: Self) -> Self {
-        let mut ret = Self::new();
-        for index in 0..N {
-            ret.array[index] = source.array[N - 1 - index].swap_bytes();
-        }
-
-        ret
-    }
-    fn from_le(source: Self) -> Self {
-        let mut ret = Self::new();
-        for index in 0..N {
-            ret.array[index] = source.array[index];
-        }
-
-        ret
-    }
-    fn to_be(self) -> Self {
-        let mut ret = Self::new();
-        for index in 0..N {
-            ret.array[index] = self.array[N - 1 - index].swap_bytes();
-        }
-
-        ret
-    }
-    fn to_le(self) -> Self {
-        let mut ret = Self::new();
-        for index in 0..N {
-            ret.array[index] = self.array[index];
-        }
-
-        ret
-    }
-    fn pow(self, n: u32) -> Self {
-        if n == 0 {
-            Self::one()
-        } else {
-            let mut ret = self;
-            for _ in 1..n {
-                ret *= self;
-            }
-            ret
-        }
-    }
-}
-
-// #endregion unimplemented
-
-// #region num-integer::Integer
-
-// Most code here from num_integer crate, unsigned implementation
-
-impl<T: MachineWord, const N: usize> num_integer::Integer for FixedUInt<T, N> {
-    fn div_floor(&self, other: &Self) -> Self {
-        *self / *other
-    }
-    fn mod_floor(&self, other: &Self) -> Self {
-        *self % *other
-    }
-    fn gcd(&self, other: &Self) -> Self {
-        // Use Stein's algorithm
-        let mut m = *self;
-        let mut n = *other;
-        let zero = Self::zero();
-        if m == zero || n == zero {
-            return m | n;
-        }
-
-        // find common factors of 2
-        let shift = (m | n).trailing_zeros();
-
-        // divide n and m by 2 until odd
-        m = m >> m.trailing_zeros();
-        n = n >> n.trailing_zeros();
-
-        while m != n {
-            if m > n {
-                m -= n;
-                m = m >> m.trailing_zeros();
-            } else {
-                n -= m;
-                n = n >> n.trailing_zeros();
-            }
-        }
-        m << shift
-    }
-    fn lcm(&self, other: &Self) -> Self {
-        if self.is_zero() && other.is_zero() {
-            return Self::zero();
-        }
-        let gcd = self.gcd(other);
-        *self * (*other / gcd)
-    }
-    fn divides(&self, other: &Self) -> bool {
-        self.is_multiple_of(other)
-    }
-    fn is_multiple_of(&self, other: &Self) -> bool {
-        (*self) % other == Self::zero()
-    }
-    fn is_even(&self) -> bool {
-        (*self) & Self::one() == Self::zero()
-    }
-    fn is_odd(&self) -> bool {
-        !self.is_even()
-    }
-    fn div_rem(&self, other: &Self) -> (Self, Self) {
-        self.div_rem(other)
-    }
-}
-
-// #endregion num-integer::Integer
-
-#[cfg(feature = "zeroize")]
-impl<T: MachineWord, const N: usize> DefaultIsZeroes for FixedUInt<T, N> {}
-
-// Helper, holds an owned copy of returned bytes
-#[derive(Eq, PartialEq, Clone, Copy, PartialOrd, Ord, Debug)]
-pub struct BytesHolder<T: MachineWord, const N: usize> {
-    array: [T; N],
-}
-
-impl<T: MachineWord, const N: usize> Default for BytesHolder<T, N> {
-    fn default() -> Self {
-        Self {
-            array: core::array::from_fn(|_| T::default()),
-        }
-    }
-}
-
-#[cfg(feature = "use-unsafe")]
-impl<T: MachineWord, const N: usize> BytesHolder<T, N> {
-    // Converts internal storage to a mutable byte slice
-    fn as_byte_slice_mut(&mut self) -> &mut [u8] {
-        // SAFETY: This is safe because the size of the array is the same as the size of the slice
-        unsafe {
-            core::slice::from_raw_parts_mut(
-                &mut self.array as *mut T as *mut u8,
-                N * core::mem::size_of::<T>(),
-            )
-        }
-    }
-    fn as_byte_slice(&self) -> &[u8] {
-        // SAFETY: This is safe because the size of the array is the same as the size of the slice
-        unsafe {
-            core::slice::from_raw_parts(
-                &self.array as *const T as *const u8,
-                N * core::mem::size_of::<T>(),
-            )
-        }
-    }
-}
-
-#[cfg(feature = "use-unsafe")]
-impl<T: MachineWord, const N: usize> Borrow<[u8]> for BytesHolder<T, N> {
-    fn borrow(&self) -> &[u8] {
-        self.as_byte_slice()
-    }
-}
-#[cfg(feature = "use-unsafe")]
-impl<T: MachineWord, const N: usize> BorrowMut<[u8]> for BytesHolder<T, N> {
-    fn borrow_mut(&mut self) -> &mut [u8] {
-        self.as_byte_slice_mut()
-    }
-}
-#[cfg(feature = "use-unsafe")]
-impl<T: MachineWord, const N: usize> AsRef<[u8]> for BytesHolder<T, N> {
-    fn as_ref(&self) -> &[u8] {
-        self.as_byte_slice()
-    }
-}
-#[cfg(feature = "use-unsafe")]
-impl<T: MachineWord, const N: usize> AsMut<[u8]> for BytesHolder<T, N> {
-    fn as_mut(&mut self) -> &mut [u8] {
-        self.as_byte_slice_mut()
-    }
-}
-#[cfg(feature = "use-unsafe")]
-impl<T: MachineWord, const N: usize> Hash for BytesHolder<T, N> {
-    fn hash<H: core::hash::Hasher>(&self, _state: &mut H) {
-        todo!()
-    }
-}
-
-#[cfg(feature = "use-unsafe")]
-impl<T: MachineWord, const N: usize> num_traits::ToBytes for FixedUInt<T, N>
-where
-    T: core::fmt::Debug,
-{
-    type Bytes = BytesHolder<T, N>;
-
-    fn to_be_bytes(&self) -> Self::Bytes {
-        let mut ret = Self::Bytes { array: self.array };
-        let _ = self.to_be_bytes(ret.as_byte_slice_mut());
-        ret
-    }
-
-    fn to_le_bytes(&self) -> Self::Bytes {
-        let mut ret = Self::Bytes { array: self.array };
-        let _ = self.to_le_bytes(ret.as_byte_slice_mut());
-        ret
-    }
-}
-
-#[cfg(feature = "use-unsafe")]
-impl<T: MachineWord, const N: usize> num_traits::FromBytes for FixedUInt<T, N>
-where
-    T: core::fmt::Debug,
-{
-    type Bytes = BytesHolder<T, N>;
-
-    fn from_be_bytes(bytes: &Self::Bytes) -> Self {
-        Self::from_be_bytes(bytes.as_ref())
-    }
-
-    fn from_le_bytes(bytes: &Self::Bytes) -> Self {
-        Self::from_le_bytes(bytes.as_ref())
-    }
-}
 
 #[cfg(test)]
 mod tests {
@@ -1789,70 +827,5 @@ mod tests {
         assert_eq!(u16_ver.to_be_bytes(&mut output_buffer).unwrap(), &be_bytes);
         assert_eq!(u32_ver.to_le_bytes(&mut output_buffer).unwrap(), &le_bytes);
         assert_eq!(u32_ver.to_be_bytes(&mut output_buffer).unwrap(), &be_bytes);
-    }
-
-    fn test_helper<T: num_traits::ToBytes>(input: &T, expected_be: &[u8]) {
-        let mut buffer = [0u8; 256];
-        buffer[..expected_be.len()].copy_from_slice(expected_be);
-        let expected_le = &mut buffer[..expected_be.len()];
-        expected_le.reverse();
-
-        let out = input.to_be_bytes();
-        assert_eq!(out.as_ref(), expected_be);
-        let out = input.to_le_bytes();
-        assert_eq!(out.as_ref(), expected_le);
-    }
-
-    #[cfg(feature = "use-unsafe")]
-    #[test]
-    fn test_to_bytes() {
-        test_helper(&0xAB_u8, &[0xAB_u8]);
-        test_helper(&0xABCD_u16, &[0xAB, 0xCD]);
-        test_helper(
-            &FixedUInt::<u8, 4>::from_u32(0x12345678).unwrap(),
-            &[0x12, 0x34, 0x56, 0x78],
-        );
-        test_helper(
-            &FixedUInt::<u16, 2>::from_u32(0x12345678).unwrap(),
-            &[0x12, 0x34, 0x56, 0x78],
-        );
-        test_helper(
-            &FixedUInt::<u32, 1>::from_u32(0x12345678).unwrap(),
-            &[0x12, 0x34, 0x56, 0x78],
-        );
-    }
-
-    fn from_helper<T>(input: &[u8], expected: T)
-    where
-        T: num_traits::FromBytes + core::fmt::Debug + core::cmp::PartialEq,
-        T::Bytes: num_traits::ops::bytes::NumBytes + Default + core::fmt::Debug,
-    {
-        let mut bytes = T::Bytes::default();
-        bytes.as_mut().copy_from_slice(input);
-        let result = T::from_be_bytes(&bytes);
-        assert_eq!(result, expected);
-        bytes.as_mut().reverse();
-        let result = T::from_le_bytes(&bytes);
-        assert_eq!(result, expected);
-    }
-
-    #[cfg(feature = "use-unsafe")]
-    #[test]
-    fn test_from_bytes() {
-        from_helper(&[0xAB_u8], 0xAB_u8);
-        from_helper(&[0xAB, 0xCD], 0xABCD_u16);
-        from_helper(&[0x12, 0x34, 0x56, 0x78], 0x12345678_u32);
-        from_helper(
-            &[0x12, 0x34, 0x56, 0x78],
-            FixedUInt::<u8, 4>::from_u32(0x12345678).unwrap(),
-        );
-        from_helper(
-            &[0x12, 0x34, 0x56, 0x78],
-            FixedUInt::<u16, 2>::from_u32(0x12345678).unwrap(),
-        );
-        from_helper(
-            &[0x12, 0x34, 0x56, 0x78],
-            FixedUInt::<u32, 1>::from_u32(0x12345678).unwrap(),
-        );
     }
 }
